compatible types

+++ get_next_line.h
@@ -24,6 +24,6 @@ char	*get_next_line(int fd);
 char	*ft_strchr(const char *s, int c);
 size_t	ft_strlen(const char *s);
 char	*ft_strdup(const char *s1);
-char	*ft_substr(char *s, int start, int len);
+char	*ft_substr(char *s, size_t start, size_t len);
 
 #endif
